+++ src/main/java/org/sustech/project/service/DataAnalysisService.java
@@ -9,7 +9,6 @@ import org.sustech.project.dto.TopicFrequencyDTO;
 import org.sustech.project.repository.AnswerRepository;
 import org.sustech.project.repository.QuestionRepository;
 
-import java.util.ArrayList;
 import java.util.List;
 import java.util.stream.Collectors;
 
@@ -42,28 +41,28 @@ public class DataAnalysisService {
     public List<MistakeDTO> getCommonMistakes() {
         List<Object[]> results = questionRepository.findCommonMistakes();
         return results.stream()
-                .map(result -> new MistakeDTO((String) result[0], (Integer) result[1]))
+                .map(result -> new MistakeDTO((String) result[0], (Long) result[1]))
                 .collect(Collectors.toList());
     }
 
     public List<AnswerQualityDTO> getAnswerQualityByTime() {
         List<Object[]> results = answerRepository.getQualityByTime();
         return results.stream()
-                .map(result -> new AnswerQualityDTO((String) result[0], (Double) result[1], (Double) result[2], (Integer) result[3]))
+                .map(result -> new AnswerQualityDTO((String) result[0], (Double) result[1], (Double) result[2], (Long) result[3]))
                 .collect(Collectors.toList());
     }
 
     public List<AnswerQualityDTO> getAnswerQualityByReputation() {
         List<Object[]> results = answerRepository.getQualityByReputation();
         return results.stream()
-                .map(result -> new AnswerQualityDTO((String) result[0], (Double) result[1], (Double) result[2], (Integer) result[3]))
+                .map(result -> new AnswerQualityDTO((String) result[0], (Double) result[1], (Double) result[2], (Long) result[3]))
                 .collect(Collectors.toList());
     }
 
     public List<AnswerQualityDTO> getAnswerQualityByLength() {
         List<Object[]> results = answerRepository.getQualityByLength();
         return results.stream()
-                .map(result -> new AnswerQualityDTO((String) result[0], (Double) result[1], (Double) result[2], (Integer) result[3]))
+                .map(result -> new AnswerQualityDTO((String) result[0], (Double) result[1], (Double) result[2], (Long) result[3]))
                 .collect(Collectors.toList());
     }
